Enable trimmable typemap replacement APIs
Share JNI remapping lookup between native and trimmable typemap type managers and enable replacement type and method lookup for TrimmableTypeMapTypeManager.

Re-enable the CoreCLRTrimmable Java.Interop replacement tests covered by the implementation.
